+++ src/Battlescape/InfoboxOKState.cpp
@@ -18,7 +18,6 @@
  */
 #include "InfoboxOKState.h"
 #include "../Engine/Game.h"
-#include "../Resource/ResourcePack.h"
 #include "../Engine/Language.h"
 #include "../Engine/Palette.h"
 #include "../Interface/TextButton.h"
