+++ src/Battlescape/InfoboxState.h
@@ -28,6 +28,7 @@ namespace OpenXcom
 class Game;
 class Text;
 class Timer;
+class Window;
 
 /**
  * Window that briefly shows an info like : Yasuaki Okamoto Has Panicked. It disappears after 2 seconds.
