+++ src/Battlescape/MiniMapState.cpp
@@ -22,15 +22,10 @@
 #include "../Engine/Game.h"
 #include "../Engine/Screen.h"
 #include "../Engine/InteractiveSurface.h"
-#include "../Engine/SurfaceSet.h"
 #include "../Resource/ResourcePack.h"
 #include "../Engine/Palette.h"
-#include "../Engine/CrossPlatform.h"
-#include "../Savegame/SavedGame.h"
-#include "../Ruleset/MapDataSet.h"
 #include "../Interface/Text.h"
 #include "MiniMapView.h"
-#include "Map.h"
 #include "Camera.h"
 #include "../Engine/Timer.h"
 #include "../Engine/Action.h"
