+++ src/Battlescape/PatrolBAIState.h
@@ -20,13 +20,14 @@
 #define OPENXCOM_PATROLBAISTATE_H
 
 #include "BattleAIState.h"
+#include "BattlescapeGame.h"
+#include <yaml-cpp/yaml.h>
 
 namespace OpenXcom
 {
 
-class BattlescapeGame;
 class BattleUnit;
-class BattlescapeState;
+class SavedBattleGame;
 class Node;
 
 /**
