+++ src/Battlescape/PromotionsState.h
@@ -20,7 +20,6 @@
 #define OPENXCOM_PROMOTIONSSTATE_H
 
 #include "../Engine/State.h"
-#include <string>
 
 namespace OpenXcom
 {
